requirements.txt
django
djangorestframework
drf-test-generator
pytest
pytest-django